fix: Ship FFmpeg libraries to prevent using those shipped with OBS.
This plugin has a dependency on FFmpeg that is shipped with the OBS installation. In OBS 30.2.0, one of the dependencies, FFmpeg, was updated. Depending on the FFmpeg shipped with OBS is not part of the public API, so it's very easy to cause issues.

Until the FFmpeg libraries when the next OBS public API breaking change occurs,